server/pyproject.toml
[tool.poetry]
name = "langchain-nextjs-fastapi"
version = "0.1.0"
description = "ChatBot Proyect"
authors = ["Your Name <m.barreiro404@proton.me>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "1.10.14"
langchain-core = "^0.1.31"
python-multipart = "^0.0.9"
fastapi = "^0.110.0"
langchain-community = "^0.0.28"
langchain-openai = "^0.0.8"
openai = "^1.13.3"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
langchain = {extras = ["docarray"], version = "^0.1.12"}


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true